Add files via upload
This Python-based project develops an interactive application for determining the most appropriate clothing size based on a user's bust, waist, and hip measurements. Implemented in a Jupyter Notebook using ipywidgets, the application provides a simple and user-friendly GUI, making it an ideal solution for online shoppers looking to accurately predict clothing sizes.